# Further optimize websocket frame mask