@crertel 2023-12-26
Oh, this is super neat!

I'm kinda surprised this works, given that we're pulling out the bulk binary xoring operations and are doing everything with one-off XORs.