@dmorneau 2023-12-27
The crypto NIF is called in batches of 20kB to allow yielding ([do_exor](https://github.com/erlang/otp/blob/2bcabe9e94d6328767729d8ff4def9bc229955ab/lib/crypto/src/crypto.erl#L2380)), so on larger inputs we might actually beat it in allocations, because the append doubles the binary size every time it's full (up to 16MB, after which it grows in 20% increments).